+++ CHANGELOG.md
@@ -3,6 +3,7 @@ next minor
 
 *   (improvement) Made `PaginatedList` generic in PhpStan.
 *   (feature) Added `PropertiesSortableHandler`.
+*   (bug) Make sortable handler more robust against cases where the doctrine entity isn't a singleton.
 
 
 7.2.3